+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- enable CSRF protection
+- Employee Module
+  - Functionality to update employee's information - ***/employee/<employee_id>/update***
+  - Functionality to inactivate/activate employees - ***/employee/<employee_id>/activate(inactivate)***
+  - More testes
+  - coerce in Employee Model to process from obj to form
+
 ### Updated
 - ***layout.html*** to provide access to different views